feat(compat-node-v15): added compatibility with node v15
