+++ .github/workflows/release.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Install node
         uses: actions/setup-node@v2-beta
         with:
-          node-version: 12.x
+          node-version: 15.x
       - name: Install Dependencies
         run: yarn install --frozen-lockfile
       - name: Lint
