+++ package.json
@@ -14,7 +14,7 @@
     "posttest": "tap --coverage-report=html --coverage-report=cobertura"
   },
   "engines": {
-    "node": ">=12.0.0 <14.0.0"
+    "node": ">=12.0.0"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^15.1.0",
